moves files to src
