+++ src/Dockerfile
@@ -8,7 +8,7 @@ WORKDIR /app
 COPY requirements.txt .
 RUN python -m pip install -r requirements.txt
 EXPOSE 8000:8000
-COPY app .
+COPY . .
 
 # Start the app
-CMD uvicorn main:app --host 0.0.0.0 --port 8000
+CMD uvicorn app.main:app --host 0.0.0.0 --port 8000
