+++ src/app/static/favicon/apple-touch-icon.png
