+++ src/app/static/favicon/favicon-32x32.png
