+++ src/app/static/favicon/site.webmanifest
